+++ github-contribution-grid@KopfdesDaemons/files/github-contribution-grid@KopfdesDaemons/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ## [2.0.0] 25.02.026
 
+- New features:
+  - Added skeleton grid for setup and error UI
+- Fixes
+  - Fixed blocking on grid rendering
 - Refactoring:
   - Added `deskletManager` as Fallback for `require` imports
   - Changed Desklet entry point to `on_desklet_removed`
@@ -14,5 +18,3 @@
   - Removed unused styles in `stylesheet,css`
   - UI logic extracted into UI Helper
   - Added Soup 2 compatibility
-- Fixes
-  - Fixed blocking on grid rendering